Move variables/domain into qililab.core, WIP implementation of new gates
