+++ src/qililab/exceptions/variable_allocated.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from qililab.qprogram.variable import Variable
+from qililab.core.variables import Variable
 
 """VariableAllocated Exception class"""
