+++ src/qililab/qprogram/blocks/loop.py
@@ -15,8 +15,8 @@
 
 import numpy as np
 
+from qililab.core.variables import Variable
 from qililab.qprogram.blocks.block import Block
-from qililab.qprogram.variable import Variable
 from qililab.yaml import yaml
