+++ src/qililab/qprogram/operations/get_parameter.py
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from qililab.core.variables import Variable
 from qililab.qprogram.operations.operation import Operation
-from qililab.qprogram.variable import Variable
 from qililab.typings.enums import Parameter
 from qililab.yaml import yaml
