+++ src/qililab/qprogram/operations/operation.py
@@ -13,8 +13,8 @@
 # limitations under the License.
 
 
+from qililab.core.variables import Variable
 from qililab.qprogram.element import Element
-from qililab.qprogram.variable import Variable
 
 
 class Operation(Element):
