+++ src/qililab/qprogram/operations/set_parameter.py
@@ -13,8 +13,8 @@
 # limitations under the License.
 
 
+from qililab.core.variables import Variable
 from qililab.qprogram.operations.operation import Operation
-from qililab.qprogram.variable import Variable
 from qililab.typings.enums import Parameter
 from qililab.yaml import yaml
