+++ src/qililab/qprogram/structured_program.py
@@ -15,9 +15,9 @@ from collections import deque
 
 import numpy as np
 
+from qililab.core.variables import Domain, FloatVariable, IntVariable, Variable
 from qililab.exceptions.variable_allocated import VariableAllocated
 from qililab.qprogram.blocks import Average, Block, ForLoop, InfiniteLoop, Loop, Parallel
-from qililab.qprogram.variable import Domain, FloatVariable, IntVariable, Variable
 from qililab.yaml import yaml
