+++ src/qililab/settings/digital/gate_event_settings.py
@@ -12,13 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from dataclasses import dataclass
+from pydantic import BaseModel, ConfigDict, field_serializer, field_validator
 
 from qililab.typings.enums import Parameter
+from qililab.utils.serialization import deserialize, serialize
+from qililab.waveforms.iq_waveform import IQWaveform
 
 
-@dataclass
-class GateEventSettings:
+class GateEventSettings(BaseModel):
     """Settings for a single gate event. A gate event is an element of a gate schedule, which is the
     sequence of gate events that define what a gate does (the pulse events it consists of).
 
@@ -30,32 +31,24 @@ class GateEventSettings:
         pulse (GatePulseSettings): settings of the bus to be launched
         wait_time (int): time to wait w.r.t gate start time (taken as 0) before launching the pulse
     """
-
-    @dataclass
-    class GatePulseSettings:
-        """Single pulse settings
-
-        Attributes:
-            amplitude (float): amplitude of the pulse
-            phase (float): phase of the pulse
-            duration (int): pulse duration
-            shape (dict): pulse envelope
-            options (dict): optional gate dependant parameters
-        """
-
-        amplitude: float
-        phase: float
-        duration: int
-        shape: dict
-        options: dict | None = None
+    model_config = ConfigDict(validate_by_name=True, validate_by_alias=True, arbitrary_types_allowed=True)
 
     bus: str
-    pulse: GatePulseSettings
+    phase: float
+    waveform: IQWaveform
+    # pulse: GatePulseSettings
     wait_time: int = 0
+    options: dict | None = None
 
-    def __post_init__(self):
-        """post init method to initialize pulse settings from runcard yaml file"""
-        self.pulse = self.GatePulseSettings(**self.pulse)
+    @field_serializer("waveform")
+    def _serialize_waveform(self, waveform: IQWaveform, _info):
+        return serialize(waveform)
+
+    @field_validator("waveform", mode="before")
+    def _load_waveform(cls, v):
+        if isinstance(v, str):
+            return deserialize(v, IQWaveform)
+        return v
 
     def set_parameter(self, parameter: Parameter, value: float | str | bool):
         """Change a given parameter from settings. Will look up into subclasses.
@@ -64,13 +57,13 @@ class GateEventSettings:
             parameter (Parameter): parameter to be set
             value (float | str | bool): value of the parameter
         """
-        param = parameter.value
-        if hasattr(self, param):
-            setattr(self, param, value)
-        elif hasattr(self.pulse, param):
-            setattr(self.pulse, param, value)
-        else:
-            self.pulse.shape[param] = value
+        # param = parameter.value
+        # if hasattr(self, param):
+        #     setattr(self, param, value)
+        # elif hasattr(self.pulse, param):
+        #     setattr(self.pulse, param, value)
+        # else:
+        #     self.pulse.shape[param] = value
 
     def get_parameter(self, parameter: Parameter):
         """Get a parameter from settings. Will look up into subclasses.
@@ -78,9 +71,9 @@ class GateEventSettings:
         Args:
             parameter (Parameter): Parameter to get.
         """
-        param = parameter.value
-        if hasattr(self, param):
-            return getattr(self, param)
-        if hasattr(self.pulse, param):
-            return getattr(self.pulse, param)
-        return self.pulse.shape[param]
+        # param = parameter.value
+        # if hasattr(self, param):
+        #     return getattr(self, param)
+        # if hasattr(self.pulse, param):
+        #     return getattr(self.pulse, param)
+        # return self.pulse.shape[param]
