+++ src/qililab/waveforms/iq_drag.py
@@ -14,8 +14,7 @@
 
 """Waveform protocol class."""
 
-from qililab.qprogram.decorators import requires_domain
-from qililab.qprogram.variable import Domain
+from qililab.core.variables import Domain, requires_domain
 from qililab.waveforms.gaussian import Gaussian
 from qililab.waveforms.gaussian_drag_correction import GaussianDragCorrection
 from qililab.waveforms.iq_waveform import IQWaveform
