+++ src/qililab/waveforms/snz.py
@@ -17,8 +17,7 @@
 import numpy as np
 
 from qililab.config import logger
-from qililab.qprogram.decorators import requires_domain
-from qililab.qprogram.variable import Domain
+from qililab.core.variables import Domain, requires_domain
 from qililab.yaml import yaml
 
 from .waveform import Waveform
